requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
opencv-python
Pillow
pytorch-fid
lpips
scikit-image
matplotlib
seaborn
plotly
gradio>=3.50.0
numpy
pandas
tqdm
PyYAML
fastapi
fastapi[all]
uvicorn
uvicorn[standard]
python-multipart
websockets
requests
python-dotenv
starlette
onnx
onnxscript
onnxruntime
onnxruntime-directml
onnxruntime-gpu
